reafactor: livesync tests
Split livesync helpers per app.
