+++ core/utils/npm.py
@@ -36,17 +36,17 @@ class Npm(object):
         assert '.tgz' in output, 'Failed to find tarball of {0} package.'.format(package)
         npm_package = output.split('/')[-1].split('\n')[0]
         src_file = os.path.join(Settings.TEST_SUT_HOME, npm_package)
-        File.clean(path=output_file)
+        File.delete(path=output_file)
         Npm.__run_npm_command('pack ' + output, folder=Settings.TEST_SUT_HOME)
         File.copy(src=src_file, target=output_file)
-        File.clean(src_file)
+        File.delete(src_file)
 
     @staticmethod
     def pack(folder, output_file):
         Npm.__run_npm_command('pack', folder=folder)
         src_file = File.find_by_extension(folder=folder, extension='tgz')[0]
         File.copy(src=src_file, target=output_file)
-        File.clean(src_file)
+        File.delete(src_file)
 
     @staticmethod
     def install(package='', option='', folder=None):
