+++ core/utils/run.py
@@ -67,7 +67,7 @@ def run(cmd, cwd=Settings.TEST_RUN_HOME, wait=True, timeout=600, fail_safe=False
 
         # noinspection PyBroadException
         try:
-            File.clean(path=log_file)
+            File.delete(path=log_file)
         except Exception:
             Log.debug('Failed to clean log file: {0}'.format(log_file))
         log_file = None
